# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Authenticates the user the based on the credentials and returns a token which will be used to authorize requests",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Authenticate a user",
                "parameters": [
                    {
                        "description": "Username",
                        "name": "username",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Password",
                        "name": "password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Student ID",
                        "name": "studentId",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/knust-server-status": {
            "get": {
                "description": "This checks which of the used KNUST servers are up or down",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "KNUST Servers"
                ],
                "summary": "Get the status of KNUST servers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.KNUSTServerStatusResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/knust-server-status/badge": {
            "get": {
                "description": "This sums up the status of the servers and returns an SVG badge as a summary from shields.io",
                "tags": [
                    "KNUST Servers"
                ],
                "summary": "Get the status of KNUST servers as a badge",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/news": {
            "get": {
                "description": "Returns the latest news available",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "News"
                ],
                "summary": "Get latest news",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.NewsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/news/{slug}": {
            "get": {
                "description": "Get the post details of the news based on the slug",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "News"
                ],
                "summary": "Get news post details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "News slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.NewsDetailsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user": {
            "get": {
                "description": "Returns personal, programme and contact user data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get User Data",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserDataResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/image/{studentId}": {
            "get": {
                "description": "Serves up the user image based on the student id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Student ID",
                        "name": "studentId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ContactUserData": {
            "type": "object",
            "properties": {
                "alt_personal_mobile": {
                    "type": "string"
                },
                "knust_mobile": {
                    "type": "string"
                },
                "personal_email": {
                    "type": "string"
                },
                "personal_mobile": {
                    "type": "string"
                },
                "postal_address": {
                    "type": "string"
                },
                "residential_address": {
                    "type": "string"
                },
                "school_email": {
                    "type": "string"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "models.KNUSTServer": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "models.KNUSTServerStatusResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "servers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.KNUSTServer"
                    }
                }
            }
        },
        "models.News": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.NewsDetails": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.NewsDetailsContent"
                    }
                },
                "date": {
                    "type": "string"
                },
                "featured_image": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "tile": {
                    "type": "string"
                }
            }
        },
        "models.NewsDetailsContent": {
            "type": "object",
            "properties": {
                "type": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "models.NewsDetailsResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "news": {
                    "$ref": "#/definitions/models.NewsDetails"
                }
            }
        },
        "models.NewsResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "news": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.News"
                    }
                }
            }
        },
        "models.PersonalUserData": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "other_names": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "religion": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.ProgrammeUserData": {
            "type": "object",
            "properties": {
                "indexNo": {
                    "type": "string"
                },
                "programme_stream": {
                    "type": "string"
                },
                "studentId": {
                    "type": "string"
                }
            }
        },
        "models.UserData": {
            "type": "object",
            "properties": {
                "contact": {
                    "$ref": "#/definitions/models.ContactUserData"
                },
                "personal": {
                    "$ref": "#/definitions/models.PersonalUserData"
                },
                "programme": {
                    "$ref": "#/definitions/models.ProgrammeUserData"
                }
            }
        },
        "models.UserDataResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "user_data": {
                    "$ref": "#/definitions/models.UserData"
                }
            }
        },
        "models.UserResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        }
    }
}